ai_services/requirements.txt
fastapi
uvicorn[standard]
pydantic
openai
python-dotenv
aiohttp
requests
numpy
pandas
